Scriptable Objects for Dialogues are now tagged as addressables. The TestDialogueManager holds these addressable references and the DialogueManager is passed the addressable reference to load it.
UniTasks are being used to handle the text animation of showing on the screen. This needs some cleaning up.

Added a custom editor script to add a button to Dialogue assets that will enter play mode. It would be nice to modify this to load straight into the asset on play.

Updated the first and second Dialogues to showcase basic font markup.